+++ src/finch/compile/lower.py
@@ -10,6 +10,7 @@ from ..algebra import Tensor, TensorFormat
 from ..codegen import NumpyBuffer
 from ..finch_assembly import AssemblyStructFormat
 from ..symbolic import Context, PostOrderDFS, PostWalk, Rewrite, ScopedDict, format
+from pprint import pprint
 
 
 class FinchTensorFormat(TensorFormat, ABC):
@@ -49,287 +50,6 @@ class FinchTensorFormat(TensorFormat, ABC):
     def unfurl(ctx, tns, ext, proto): ...
 
 
-class BufferizedNDArray(Tensor):
-    def __init__(self, arr: np.ndarray):
-        itemsize = arr.dtype.itemsize
-        for stride in arr.strides:
-            if stride % itemsize != 0:
-                raise ValueError("Array must be aligned to multiple of itemsize")
-        self.strides = [stride // itemsize for stride in arr.strides]
-        self._shape = arr.shape
-        self.buf = NumpyBuffer(
-            np.lib.stride_tricks.as_strided(
-                arr,
-                shape=(np.dot(arr.strides, arr.shape) // itemsize,),
-                strides=(itemsize,),
-            )
-        )
-
-    def to_numpy(self):
-        """
-        Convert the bufferized NDArray to a NumPy array.
-        This is used to get the underlying NumPy array from the bufferized NDArray.
-        """
-        itemsize = self.buf.arr.dtype.itemsize
-        return np.lib.stride_tricks.as_strided(
-            self.buf.arr,
-            shape=self._shape,
-            strides=(stride * itemsize for stride in self.strides),
-        )
-
-    @property
-    def format(self):
-        """
-        Returns the format of the buffer, which is a BufferizedNDArrayFormat.
-        """
-        return BufferizedNDArrayFormat(format(self.buf), len(self.strides))
-
-    @property
-    def shape(self):
-        return self._shape
-
-    def declare(self, init, op, shape):
-        """
-        Declare a bufferized NDArray with the given initialization value,
-        operation, and shape.
-        """
-        for dim, size in zip(shape, self._shape, strict=False):
-            if dim.start != 0:
-                raise ValueError(
-                    f"Invalid dimension start value {dim.start} for ndarray"
-                    f" declaration."
-                )
-            if dim.end != size:
-                raise ValueError(
-                    f"Invalid dimension end value {dim.end} for ndarray declaration."
-                )
-        shape = tuple(dim.end for dim in shape)
-        for i in range(self.buf.length()):
-            self.buf.store(i, init)
-        return self
-
-    def freeze(self, op):
-        return self
-
-    def thaw(self, op):
-        return self
-
-    def access(self, indices, op):
-        return BufferizedNDArrayAccessor(self).access(indices, op)
-
-    def __getitem__(self, index):
-        """
-        Get an item from the bufferized NDArray.
-        This allows for indexing into the bufferized array.
-        """
-        if isinstance(index, tuple):
-            index = np.dot(index, self.strides)
-        return self.buf.load(index)
-
-    def __setitem__(self, index, value):
-        """
-        Set an item in the bufferized NDArray.
-        This allows for indexing into the bufferized array.
-        """
-        if isinstance(index, tuple):
-            index = np.ravel_multi_index(index, self._shape)
-        self.buf.store(index, value)
-
-
-class BufferizedNDArrayFormat(FinchTensorFormat, AssemblyStructFormat):
-    """
-    A format for bufferized NumPy arrays that provides metadata about the array.
-    This includes the fill value, element type, and shape type.
-    """
-
-    @property
-    def struct_name(self):
-        return "BufferizedNDArray"
-
-    @property
-    def struct_fields(self):
-        return [
-            ("buf", self.buf),
-            ("_ndim", self._ndim),
-        ]
-
-    def __init__(self, buf, ndim: int):
-        self.buf = buf
-        self._ndim = ndim
-
-    def __eq__(self, other):
-        if not isinstance(other, BufferizedNDArrayFormat):
-            return False
-        return self.buf == other.buf and self._ndim == other._ndim
-
-    def __hash__(self):
-        return hash((self.buf, self._ndim))
-
-    @property
-    def ndim(self) -> int:
-        return self._ndim
-
-    @property
-    def fill_value(self) -> Any:
-        return np.zeros((), dtype=self.buf.dtype)[()]
-
-    @property
-    def element_type(self):
-        return self.buf.dtype.type
-
-    @property
-    def shape_type(self) -> tuple:
-        return tuple(np.int_ for _ in range(self._ndim))
-
-    def lower_declare(self, ctx, tns, init, op, shape):
-        i_var = asm.Variable("i", self.buf.length_type)
-        buf = asm.Stack(tns.obj.buf, self.buf)
-        body = asm.Store(
-            buf,
-            i_var,
-            asm.Literal(init.val),
-        )
-        ctx.exec(asm.ForLoop(i_var, asm.Literal(0), asm.Length(buf), body))
-        return
-
-    def lower_freeze(self, ctx, tns, op):
-        return tns
-
-    def lower_thaw(self, ctx, tns, op):
-        return tns
-
-    def unfurl(self, ctx, tns, ext, mode, proto):
-        acc_t = BufferizedNDArrayAccessorFormat(self, 0, self.buf.length_type(0), None)
-        return acc_t.unfurl(ctx, tns, ext, mode, proto)
-
-    def lower_unwrap(self, ctx, obj): ...
-
-    def lower_increment(self, ctx, obj, val): ...
-
-    def asm_unpack(self, ctx, var_n, val):
-        """
-        Unpack the into asm context.
-        """
-        stride = []
-        for i in range(self._ndim):
-            stride_i = asm.Variable(f"{var_n}_stride_{i}", self.buf.length_type)
-            stride.append(stride_i)
-            stride_e = asm.GetAttr(val, "stride")
-            stride_i_e = asm.GetAttr(stride_e, f"element_{i}")
-            ctx.exec(asm.Assign(stride_i, stride_i_e))
-        buf = asm.Variable(f"{var_n}_buf", self.buf)
-        buf_e = asm.GetAttr(val, "buf")
-        ctx.exec(asm.Assign(buf, buf_e))
-        buf_s = asm.Slot(f"{var_n}_buf", self.buf)
-        ctx.exec(asm.Unpack(buf_s, buf))
-
-        class BufferizedNDArrayFields(NamedTuple):
-            stride: list[asm.Variable]
-            buf: asm.Variable
-            buf_s: asm.Slot
-
-        return BufferizedNDArrayFields(stride, buf, buf_s)
-
-    def asm_repack(self, ctx, lhs, obj):
-        """
-        Repack the buffer from C context.
-        """
-        ctx.exec(asm.Repack(obj.buf))
-        return
-
-
-class BufferizedNDArrayAccessor(Tensor):
-    """
-    A class representing a tensor view that is bufferized.
-    This is used to create a view of a tensor with a specific extent.
-    """
-
-    def __init__(self, tns: BufferizedNDArray, nind=None, pos=None, op=None):
-        self.tns = tns
-        if pos is None:
-            pos = format(self.tns).buf.length_type(0)
-        self.pos = pos
-        self.op = op
-        if nind is None:
-            nind = 0
-        self.nind = nind
-
-    @property
-    def format(self):
-        return BufferizedNDArrayAccessorFormat(
-            format(self.tns), self.nind, format(self.pos), self.op
-        )
-
-    @property
-    def shape(self):
-        return self.tns.shape[self.nind :]
-
-    def access(self, indices, op):
-        pos = self.pos + np.dot(
-            indices, self.tns.strides[self.nind : self.nind + len(indices)]
-        )
-        return BufferizedNDArrayAccessor(self.tns, self.nind + len(indices), pos, op)
-
-    def unwrap(self):
-        """
-        Unwrap the tensor view to get the underlying tensor.
-        This is used to get the original tensor from a tensor view.
-        """
-        assert self.ndim == 0, "Cannot unwrap a tensor view with non-zero dimension."
-        return self.tns.buf.load(self.pos)
-
-    def increment(self, val):
-        """
-        Increment the tensor view with a value.
-        This updates the tensor at the specified index with the operation and value.
-        """
-        if self.op is None:
-            raise ValueError("No operation defined for increment.")
-        assert self.ndim == 0, "Cannot unwrap a tensor view with non-zero dimension."
-        self.tns.buf.store(self.pos, self.op(self.tns.buf.load(self.pos), val))
-        return self
-
-
-class BufferizedNDArrayAccessorFormat(FinchTensorFormat):
-    def __init__(self, tns, nind, pos, op):
-        self.tns = tns
-        self.nind = nind
-        self.pos = pos
-        self.op = op
-
-    def __eq__(self, other):
-        return (
-            isinstance(other, BufferizedNDArrayAccessorFormat)
-            and self.tns == other.tns
-            and self.nind == other.nind
-            and self.pos == other.pos
-            and self.op == other.op
-        )
-
-    def __hash__(self):
-        return hash((self.tns, self.nind, self.pos, self.op))
-
-    @property
-    def ndim(self) -> int:
-        return self.tns.ndim - self.nind
-
-    @property
-    def shape_type(self) -> tuple:
-        return self.tns.shape_type[self.nind :]
-
-    @property
-    def fill_value(self) -> Any:
-        return self.tns.fill_value
-
-    @property
-    def element_type(self):
-        return self.tns.element_type
-
-    def lower_unwrap(self, ctx, obj): ...
-
-    def lower_increment(self, ctx, obj, val): ...
-
-    def unfurl(self, ctx, tns, ext, mode, proto): ...
 
 
 @dataclass(eq=True, frozen=True)
@@ -368,6 +88,12 @@ def dimension(tns, mode):
 class ExtentFormat:
     start: Any
     end: Any
+
+    def get_start(self, ext):
+        return asm.GetAttr(ext, "start")
+
+    def get_end(self, ext):
+        return asm.GetAttr(ext, "end")
 
     def lower_loop(self, ctx, idx, ext, body):
         """
@@ -666,9 +392,7 @@ class LoopletPass(ABC):
     @abstractmethod
     def priority(self): ...
 
-    def lt(self, other):
-        if other is None:
-            return False
+    def __lt__(self, other):
         assert isinstance(other, LoopletPass)
         return self.priority < other.priority
 
@@ -678,34 +402,12 @@ class DefaultPass(LoopletPass):
     def priority(self):
         return float("-inf")
 
-
-class LookupPass(LoopletPass):
-    @property
-    def priority(self):
-        return 0
-
     def __call__(self, ctx, idx, ext, body):
-        idx_2 = asm.Variable(self.freshen("i"), idx.result_format)
+        """
+        Default pass that does nothing. This is used when no other pass is selected.
+        """
+        assert False
 
-        def lookup_node(node):
-            match node:
-                case ntn.Access(tns, mode, (j, *idxs)):
-                    if j == idx:
-                        tns = ctx.resolve(tns)
-                        tns_2 = tns.result_format.lookup(
-                            ctx,
-                            tns,
-                            idx_2,
-                        )
-                        return ntn.Access(tns_2, mode, (j, *idxs))
-            return None
-
-        body_2 = PostWalk(lookup_node)(body)
-        ctx.exec(
-            asm.ForLoop(
-                idx_2, asm.Literal(ext.start), asm.Literal(ext.end), ctx(body_2)
-            )
-        )
 
 
 class LoopletContext(Context):
@@ -733,12 +435,13 @@ class LoopletContext(Context):
         return self.ctx.emit()
 
     def select_pass(self, body):
+        pprint(body)
         def pass_request(node):
             match node:
                 case ntn.Access(tns, _, (j, *_)):
                     if j == self.idx:
-                        return tns.pass_request()
-            return None
+                        return tns.pass_request
+            return DefaultPass()
 
         return max(map(pass_request, PostOrderDFS(body)))
